+++ src/roiextractors/extractors/memmapextractors/memmapextractors.py
@@ -57,7 +57,7 @@ class MemmapImagingExtractor(ImagingExtractor):
             )
 
         if frame_idxs is None:
-            frame_idxs = [frame for frame in range(self.get_num_frames())]
+            frame_idxs = [frame for frame in range(self.get_num_samples())]
 
         frames = self._video.take(indices=frame_idxs, axis=0)
         if channel is not None: