+++ tests/test_internals/test_testing_tools.py
@@ -37,8 +37,10 @@ class TestDummySegmentationExtractor(TestCase):
 
         # Test frame_to_time
         times = np.arange(self.num_frames) / self.sampling_frequency
-        assert_array_equal(segmentation_extractor.frame_to_time(frames=np.arange(self.num_frames)), times)
-        self.assertEqual(segmentation_extractor.frame_to_time(frames=8), times[8])
+        assert_array_equal(
+            segmentation_extractor.sample_indices_to_time(sample_indices=np.arange(self.num_frames)), times
+        )
+        self.assertEqual(segmentation_extractor.sample_indices_to_time(sample_indices=8), times[8])
 
         # Test image masks
         assert segmentation_extractor.get_roi_image_masks().shape == (self.num_rows, self.num_columns, self.num_rois)
@@ -122,8 +124,8 @@ class TestDummySegmentationExtractor(TestCase):
         times = np.arange(self.num_frames) / self.sampling_frequency
         segmentation_extractor._times = times
 
-        self.assertEqual(segmentation_extractor.frame_to_time(frames=2), times[2])
+        self.assertEqual(segmentation_extractor.sample_indices_to_time(sample_indices=2), times[2])
         assert_array_equal(
-            segmentation_extractor.frame_to_time(frames=np.arange(self.num_frames)),
+            segmentation_extractor.sample_indices_to_time(sample_indices=np.arange(self.num_frames)),
             times,
         )